+++ src/content-scripts/page-adapters/gemini.ts
@@ -113,6 +113,8 @@ const geminiPageAdapter = async () => {
             PageLive.KeybindManager.Keybinds.ToggleSidebar,
             toggleSidebar
         );
+        // Add 'more' keybinds to page info dialog. These are native keybinds / shortcuts that are not provided by PageLive, but are still useful for users to know.
+        pl.pageInfoDialog.addMoreKeybind("Ctrl + K", "Open dialog to search chat");
 
         // Add callback to be executed the next time dialog is shown
         pl.pageInfoDialog.onEveryOpenCallback = onDialogOpen